enforce calling Mapper.Initialize once
