+++ src/AutoMapper/Mapper.cs
@@ -8,6 +8,7 @@ namespace AutoMapper
     public class Mapper : IRuntimeMapper
     {
         private const string InvalidOperationMessage = "Mapper not initialized. Call Initialize with appropriate configuration. If you are trying to use mapper instances through a container or otherwise, make sure you do not have any calls to the static Mapper.Map methods, and if you're using ProjectTo or UseAsDataSource extension methods, make sure you pass in the appropriate IConfigurationProvider instance.";
+        private const string AlreadyInitialized = "Mapper already initialized. You must call Initialize once per application domain/process.";
 
         #region Static API
 
@@ -29,7 +30,7 @@ namespace AutoMapper
         public static IMapper Instance
         {
             get => _instance ?? throw new InvalidOperationException(InvalidOperationMessage);
-            private set => _instance = value;
+            private set => _instance = (_instance == null) ? value : throw new InvalidOperationException(AlreadyInitialized);
         }
 
         /// <summary>
