+++ src/UnitTests/Projection/ProjectionTests.cs
@@ -14,7 +14,7 @@
 
         [Fact]
         public void Direct_assignability_shouldnt_trump_custom_projection() {
-            Mapper.Initialize(x => {
+            var config = new MapperConfiguration(x => {
                 x.CreateMap<string, string>()
                     .ProjectUsing(s => _niceGreeting);
 
@@ -24,7 +24,7 @@
 
             var target = new[] { new Source() { Greeting = _badGreeting } }
                             .AsQueryable()
-                            .ProjectTo<Target>()
+                            .ProjectTo<Target>(config)
                             .First();
 
             target.Greeting.ShouldBe(_niceGreeting);
@@ -33,14 +33,14 @@
 
         [Fact]
         public void Root_is_subject_to_custom_projection() {
-            Mapper.Initialize(x => {
+            var config = new MapperConfiguration(x => {
                 x.CreateMap<Source, Target>()
                     .ProjectUsing(s => new Target() { Greeting = _niceGreeting });
             });
 
             var target = new[] { new Source() }
                             .AsQueryable()
-                            .ProjectTo<Target>()
+                            .ProjectTo<Target>(config)
                             .First();
 
             target.Greeting.ShouldBe(_niceGreeting);
@@ -49,7 +49,7 @@
 
         [Fact]
         public void Child_nodes_are_subject_to_custom_projection() {
-            Mapper.Initialize(x => {
+            var config = new MapperConfiguration(x => {
                 x.CreateMap<SourceChild, TargetChild>()
                     .ProjectUsing(s => new TargetChild() { Greeting = _niceGreeting });
 
@@ -58,7 +58,7 @@
 
             var target = new[] { new Source() }
                             .AsQueryable()
-                            .ProjectTo<Target>()
+                            .ProjectTo<Target>(config)
                             .First();
 
             target.Child.Greeting.ShouldBe(_niceGreeting);
